chore: move err boundary to separate file
