+++ examples/SampleApp/App.tsx
@@ -20,8 +20,6 @@ import {
   OverlayProvider,
   setupCommandUIMiddlewares,
   SqliteClient,
-  SqliteClientError,
-  type SqliteClientErrorCode,
   Streami18n,
   ThemeProvider,
   useOverlayContext,
@@ -29,6 +27,7 @@ import {
 } from 'stream-chat-react-native';
 
 import { MenuDrawer } from './src/components/MenuDrawer';
+import { OfflineDbBoundary } from './src/components/OfflineDbBoundary';
 import { useSampleAppComponentOverrides } from './src/components/SampleAppComponentOverrides';
 import {
   MessageInputFloatingConfigItem,
@@ -331,71 +330,6 @@ const DrawerNavigator: React.FC = () => (
 );
 
 const isMessageAIGenerated = (message: LocalMessage) => !!message.ai_generated;
-
-/**
- * `<Chat>` throws a {@link SqliteClientError} from render when it cannot open the
- * offline database - most often `OFFLINE_DB_UNREADABLE`, meaning the file on disk
- * cannot be read (corruption, or a database left behind from a different encryption
- * mode). It never silently continues without the cache; recovery is the application's
- * decision.
- *
- * The recommended recovery, shown here: the contents are a cache, so delete the
- * database and let it rebuild from the server. The only real loss is actions that were
- * queued while offline, so a real app may want to confirm with the user first.
- *
- * The `onGiveUp` path covers the codes that mean "no usable encryption key"
- * (`SQLCIPHER_BUILD_MISSING`, `ENCRYPTION_KEY_UNAVAILABLE`). Those only occur when
- * `<Chat>` is given a `getEncryptionKey` prop, which this sample does not do - a new
- * database would then be written in plaintext, so running online-only is the safe
- * response.
- */
-type BoundaryProps = React.PropsWithChildren<{
-  onGiveUp: () => void;
-  onRetry: () => void;
-}>;
-
-type BoundaryState = { code?: SqliteClientErrorCode };
-
-class OfflineDbBoundary extends React.Component<BoundaryProps, BoundaryState> {
-  state: BoundaryState = {};
-
-  // Must return state, and render() must stop rendering the failing subtree. Returning
-  // null here would re-render the same children, they would throw again, and React
-  // would give up and unmount the whole app.
-  static getDerivedStateFromError(error: unknown) {
-    const code = (error as SqliteClientError | undefined)?.code;
-    if (!code) {
-      throw error;
-    }
-    return { code };
-  }
-
-  componentDidCatch(error: unknown) {
-    // Discriminated on `code` rather than `instanceof`: a string comparison cannot be
-    // defeated by two copies of the class ending up in one bundle.
-    const code = (error as SqliteClientError | undefined)?.code;
-
-    if (code === 'OFFLINE_DB_UNREADABLE') {
-      // The recommended recovery: the contents are a cache, so drop the database and
-      // let it rebuild. Only actions queued while offline are lost.
-      try {
-        SqliteClient.deleteDatabase();
-      } catch (deleteError) {
-        console.warn('[SampleApp] could not delete the offline database', deleteError);
-      }
-      this.props.onRetry();
-      return;
-    }
-
-    // No usable key, so a new database would be plaintext. Run online-only instead.
-    console.warn(`[SampleApp] offline encryption unavailable (${code}); going online-only`);
-    this.props.onGiveUp();
-  }
-
-  render() {
-    return this.state.code ? null : this.props.children;
-  }
-}
 
 const DrawerNavigatorWrapper: React.FC<{
   chatClient: StreamChat;
